Moved Yield to utils.rs.
