+++ components/test/src/utils.rs
@@ -1,8 +1,9 @@
 use std::path::PathBuf;
+use std::pin::Pin;
 use std::collections::HashMap;
 
-use futures::task::{Spawn, SpawnExt};
-use futures::{future, FutureExt, TryFutureExt};
+use futures::task::{Spawn, SpawnExt, LocalWaker};
+use futures::{future, Future, FutureExt, TryFutureExt, Poll};
 
 use crypto::identity::{PublicKey, Identity, 
     generate_pkcs8_key_pair, SoftwareEd25519Identity};
@@ -47,6 +48,31 @@ const CONN_TIMEOUT_TICKS: usize = 0x8;
 /// Maximum amount of concurrent applications
 /// going through the incoming connection transform at the same time
 const MAX_CONCURRENT_INCOMING_APPS: usize = 0x8;
+
+// Based on:
+// - https://rust-lang-nursery.github.io/futures-api-docs/0.3.0-alpha.13/src/futures_test/future/pending_once.rs.html#14-17
+// - https://github.com/rust-lang-nursery/futures-rs/issues/869
+pub struct Yield(usize);
+
+impl Yield {
+    pub fn new(num_yields: usize) -> Self {
+        Yield(num_yields)
+    }
+}
+
+impl Future for Yield {
+    type Output = ();
+    fn poll(mut self: Pin<&mut Self>, waker: &LocalWaker) -> Poll<Self::Output> {
+        let count = &mut self.as_mut().0;
+        *count = count.saturating_sub(1);
+        if *count == 0 {
+            Poll::Ready(())
+        } else {
+            waker.wake();
+            Poll::Pending
+        }
+    }
+}
 
 fn gen_identity<R>(rng: &R) -> impl Identity 
 where
